+++ commands/admin/delete.js
