+++ commands/api/animagine.js
