+++ commands/api/play.js
