+++ commands/donos/seradmin.js
@@ -1,4 +1,4 @@
-const { donos } = require("../database/models/donos");
+const { donos } = require("../../database/models/donos");
 
 module.exports = {
   name: "seradmin",
