+++ commands/geral/menu.js
@@ -1,4 +1,4 @@
-const menu = require("../utils/menu");
+const menu = require("../../utils/menu");
 const path = require("path")
 
 
@@ -17,7 +17,7 @@ module.exports = {
     
     try {
       
-      await sock.sendMessage(from, {image: {url: path.join(__dirname, "../assets/images/yuki.jpg")}, caption: menu(msg), mentions: [sender]}, {quoted: msg});
+      await sock.sendMessage(from, {image: {url: path.join(__dirname, "../../assets/images/yuki.jpg")}, caption: menu(msg), mentions: [sender]}, {quoted: msg});
       
       
     }
