+++ commands/geral/meulid.js
