+++ commands/geral/toimg.js
